docs(hero): update structure and move to patterns
