+++ .storybook/main.js
@@ -74,7 +74,8 @@ module.exports = {
     '../projects/canopy/src/lib/forms/date/docs/guide.stories.mdx',
     '../projects/canopy/src/lib/forms/date/docs/date-field.stories.ts',
     '../projects/canopy/src/lib/forms/sort-code/sort-code.stories.ts',
-    '../projects/canopy/src/lib/hero/hero.stories.ts',
+    '../projects/canopy/src/lib/hero/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/hero/docs/hero.stories.ts',
     '../projects/canopy/src/lib/hero-img/hero-img.stories.ts',
     '../projects/canopy/src/lib/side-nav/side-nav.stories.ts',
     // Templates
